[MNT] Move SubplotParams from figure to gridspec
Logically, this belongs more to the gridspec layout than
to the figure.

There is no API breakage involved, because
`matplotlib.figure` reimports `SubplotParams`. Do we
nevertheless want an API change note?
