+++ galleries/examples/subplots_axes_and_figures/auto_subplots_adjust.py
@@ -83,5 +83,5 @@ plt.show()
 #    - `matplotlib.transforms.BboxBase.union`
 #    - `matplotlib.transforms.Transform.inverted`
 #    - `matplotlib.figure.Figure.subplots_adjust`
-#    - `matplotlib.figure.SubplotParams`
+#    - `matplotlib.gridspec.SubplotParams`
 #    - `matplotlib.backend_bases.FigureCanvasBase.mpl_connect`
