+++ lib/matplotlib/gridspec.pyi
@@ -5,7 +5,7 @@ import numpy as np
 
 from matplotlib.axes import Axes, SubplotBase
 from matplotlib.backend_bases import RendererBase
-from matplotlib.figure import Figure, SubplotParams
+from matplotlib.figure import Figure
 from matplotlib.transforms import Bbox
 
 class GridSpecBase:
@@ -132,3 +132,29 @@ class SubplotSpec:
     def subgridspec(
         self, nrows: int, ncols: int, **kwargs
     ) -> GridSpecFromSubplotSpec: ...
+
+class SubplotParams:
+    def __init__(
+        self,
+        left: float | None = ...,
+        bottom: float | None = ...,
+        right: float | None = ...,
+        top: float | None = ...,
+        wspace: float | None = ...,
+        hspace: float | None = ...,
+    ) -> None: ...
+    left: float
+    right: float
+    bottom: float
+    top: float
+    wspace: float
+    hspace: float
+    def update(
+        self,
+        left: float | None = ...,
+        bottom: float | None = ...,
+        right: float | None = ...,
+        top: float | None = ...,
+        wspace: float | None = ...,
+        hspace: float | None = ...,
+    ) -> None: ...
